telegram-bot/requirements.txt
python-telegram-bot[job-queue]==21.6
redis[hiredis]==5.0.8
httpx==0.27.2
python-dotenv==1.0.1
